Move process_content_sections tests to new file.
